+++ ServerSide/src/database/playerinfo/PlayerModel.java
@@ -21,6 +21,8 @@ public interface PlayerModel {
 
              try {
                     db.startConnection();
+                    if(db.getConnection() == null )
+                        return null ;
                     db.setStatement(db.getConnection().createStatement()) ;
                     db.setQuerystr("select username from players where pid= "+_pid);
                     db.setResultSet(db.getStatement().executeQuery(db.getQuerystr()));  
@@ -46,6 +48,8 @@ public interface PlayerModel {
     static boolean insertRecord(String _username , String _passwd , String _email  , String _status ,long _score , String _avatar){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return false ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 
                 int checkNew=db.getStatement().executeUpdate("INSERT INTO players ( username, passwd, email, status, score, avatar) VALUES( '"+_username+"', '"+_passwd+"', '"+_email+"', '"+_status+"', "+_score+", '"+_avatar+"' )"); 
@@ -64,6 +68,8 @@ public interface PlayerModel {
     static boolean updateIdRecord(long _pid ,String _username , String _passwd , String _email  , String _status ,long _score , String _avatar){
          try {
                 db.startConnection();
+                 if(db.getConnection() == null )
+                        return false ;
                 db.setStatement(db.getConnection().createStatement()) ; 
 
                 int checkUpdate=db.getStatement().executeUpdate("UPDATE players SET username= '"+_username+"' , passwd= '"+_passwd+"' , email= '"+_email+"' , status='"+_status+"' , score= "+_score+" ,avatar= '"+_avatar+"' WHERE pid = "+_pid); 
@@ -83,6 +89,8 @@ public interface PlayerModel {
     static boolean updateUsrRecord(String _username , String _passwd , String _email  , String _status ,long _score , String _avatar){
          try {
                 db.startConnection();
+                 if(db.getConnection() == null )
+                        return false ;
                 db.setStatement(db.getConnection().createStatement()) ; 
 
                 int checkUpdate=db.getStatement().executeUpdate("UPDATE players SET passwd= '"+_passwd+"' , email= '"+_email+"' , status='"+_status+"' , score= "+_score+" ,avatar= '"+_avatar+"' WHERE username = '"+_username+"'"); 
@@ -102,6 +110,8 @@ public interface PlayerModel {
     static boolean updateUsrFieldStatus(String _username , String _status ){
          try {
                 db.startConnection();
+                 if(db.getConnection() == null )
+                        return false ;
                 db.setStatement(db.getConnection().createStatement()) ; 
 
                 int checkUpdate=db.getStatement().executeUpdate("UPDATE players SET status='"+_status+"' WHERE username = '"+_username+"'"); 
@@ -119,6 +129,8 @@ public interface PlayerModel {
     static boolean updateFieldStatus(String _status){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return false ;
                 db.setStatement(db.getConnection().createStatement()) ; 
 
                 int checkUpdate=db.getStatement().executeUpdate("UPDATE players SET status='"+_status+"'"); 
@@ -137,6 +149,8 @@ public interface PlayerModel {
     static boolean updateUsrFieldScore(String _username , long _score ){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return false ;
                 db.setStatement(db.getConnection().createStatement()) ; 
 
                 int checkUpdate=db.getStatement().executeUpdate("UPDATE players SET score="+_score+" WHERE username = '"+_username+"'"); 
@@ -157,6 +171,8 @@ public interface PlayerModel {
     static boolean updateUsrPassRecord(String _username , String _passwd , String _email  , String _status ,long _score , String _avatar){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return false ;
                 db.setStatement(db.getConnection().createStatement()) ; 
 
                 int checkUpdate=db.getStatement().executeUpdate("UPDATE players SET email= '"+_email+"' , status='"+_status+"' , score= "+_score+" ,avatar= '"+_avatar+"' WHERE username = '"+_username+"' and passwd= '"+_passwd+"'"); 
@@ -174,6 +190,8 @@ public interface PlayerModel {
     static boolean updateUsrPassFieldStatus(String _username , String _passwd , String _status ){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return false ;
                 db.setStatement(db.getConnection().createStatement()) ; 
 
                 int checkUpdate=db.getStatement().executeUpdate("UPDATE players SET status='"+_status+"' WHERE username = '"+_username+"' and passwd= '"+_passwd+"'"); 
@@ -192,6 +210,8 @@ public interface PlayerModel {
     static boolean updateUsrPassFieldScore(String _username , String _passwd ,long _score ){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return false ;
                 db.setStatement(db.getConnection().createStatement()) ; 
 
                 int checkUpdate=db.getStatement().executeUpdate("UPDATE players SET score='"+_score+"' WHERE username = '"+_username+"' and passwd= '"+_passwd+"'"); 
@@ -210,6 +230,8 @@ public interface PlayerModel {
     static boolean deleteIdRecord(long _pid){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return false ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 
                 int checkDelete=db.getStatement().executeUpdate("DELETE FROM players WHERE pid="+_pid); 
@@ -231,6 +253,8 @@ public interface PlayerModel {
     static boolean deleteUsrRecord(String _username){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return false ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 int checkDelete=db.getStatement().executeUpdate("DELETE FROM players WHERE username= '"+_username+"'"); 
                 db.endStatConnection();  // for statment with no resultset
@@ -248,6 +272,8 @@ public interface PlayerModel {
     static boolean deleteMailRecord(String _email){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return false ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 int checkDelete=db.getStatement().executeUpdate("DELETE FROM players WHERE email= '"+_email+"'"); 
                 db.endStatConnection();  // for statment with no resultset
@@ -265,6 +291,8 @@ public interface PlayerModel {
     static boolean deleteMailPassRecord(String _email, String _passwd){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return false ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 int checkDelete=db.getStatement().executeUpdate("DELETE FROM players WHERE email= '"+_email+"' and passwd= '"+_passwd+"'"); 
                 db.endStatConnection();  // for statment with no resultset
@@ -281,6 +309,8 @@ public interface PlayerModel {
     static boolean deleteUsrPassRecord(String _username, String _passwd){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return false ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 int checkDelete=db.getStatement().executeUpdate("DELETE FROM players WHERE email= '"+_username+"' and passwd= '"+_passwd+"'"); 
                 db.endStatConnection();  // for statment with no resultset
@@ -300,6 +330,8 @@ public interface PlayerModel {
     static Long selectScoreWhereUsr(String _username ){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return null ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 db.setQuerystr("select score from players where username= '"+_username+"'");
                 db.setResultSet(db.getStatement().executeQuery(db.getQuerystr()));  
@@ -324,6 +356,8 @@ public interface PlayerModel {
     static Long selectIdWhereUsr(String _username ){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return null ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 db.setQuerystr("select pid from players where username= '"+_username+"'");
                 db.setResultSet(db.getStatement().executeQuery(db.getQuerystr()));  
@@ -346,6 +380,8 @@ public interface PlayerModel {
     static String selectStatusWhereUsr(String _username ){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return null ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 db.setQuerystr("select status from players where username= '"+_username+"'");
                 db.setResultSet(db.getStatement().executeQuery(db.getQuerystr()));  
@@ -371,6 +407,8 @@ public interface PlayerModel {
     static String selectPassWhereUsr(String _username ){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return null ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 db.setQuerystr("select passwd from players where username= '"+_username+"'");
                 db.setResultSet(db.getStatement().executeQuery(db.getQuerystr()));  
@@ -394,6 +432,8 @@ public interface PlayerModel {
      static boolean selectWhereUsrPass(String _username ,String _passwd ){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return false ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 db.setQuerystr("select * from players where username= '"+_username+"' and passwd = '"+_passwd+"'");
                 db.setResultSet(db.getStatement().executeQuery(db.getQuerystr()));  
@@ -416,6 +456,8 @@ public interface PlayerModel {
     static Player selectPlayerWhereUsrPass(String _username ,String _passwd ){
         try {
             db.startConnection();
+            if(db.getConnection() == null )
+                        return null ;
             db.setStatement(db.getConnection().createStatement()) ;
             db.setQuerystr("select * from players where username= '"+_username+"' and passwd = '"+_passwd+"'");
             db.setResultSet(db.getStatement().executeQuery(db.getQuerystr()));  
@@ -442,6 +484,8 @@ public interface PlayerModel {
     static String selectMailWhereUsr(String _username ){
         try {
             db.startConnection();
+            if(db.getConnection() == null )
+                        return null ;
             db.setStatement(db.getConnection().createStatement()) ;
             db.setQuerystr("select email from players where username= '"+_username+"'");
             db.setResultSet(db.getStatement().executeQuery(db.getQuerystr()));  
@@ -465,6 +509,8 @@ public interface PlayerModel {
      static Long selectScoreWhereUsrPass(String _username ,String _passwd ){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return null ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 db.setQuerystr("select score from players where username= '"+_username+"' and passwd='"+_passwd+"'");
                 db.setResultSet(db.getStatement().executeQuery(db.getQuerystr()));  
@@ -490,6 +536,8 @@ public interface PlayerModel {
       static String selectStatusWhereUsrPass(String _username ,String _passwd ){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return null ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 db.setQuerystr("select status from players where username= '"+_username+"' and passwd='"+_passwd+"'");
   
@@ -514,6 +562,8 @@ public interface PlayerModel {
     static Vector<Player> selectAllWhereStatus(String _status ){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return null ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 db.setQuerystr("select * from players where status= '"+_status+"'");
   
@@ -543,6 +593,8 @@ public interface PlayerModel {
      static Vector<Player> selectAllPlayers(){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return null ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 db.setQuerystr("select * from players ");
   
@@ -570,6 +622,8 @@ public interface PlayerModel {
      static Vector<Player> selectAllPlayersOrderByDESC(String colName ){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return null ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 db.setQuerystr("select * from players ORDER BY "+colName+" DESC ");
   
@@ -598,6 +652,8 @@ public interface PlayerModel {
      static Vector<Player> selectAllPlayersOrderByASC(String colName ){
          try {
                 db.startConnection();
+                if(db.getConnection() == null )
+                        return null ;
                 db.setStatement(db.getConnection().createStatement()) ;
                 db.setQuerystr("select * from players ORDER BY "+colName+" ASC ");
